fixed name_to_properties and its testing
