+++ pytest.ini
@@ -6,4 +6,4 @@ filterwarnings =
 markers =
     slow: marks tests as slow (deselect with '-m "not slow"')
 
-addopts = --cov=gcms_data_analysis --cov-report=html:docs/_coverage_html
+addopts = -s
