+++ tests/conftest.py
@@ -5,7 +5,49 @@ import pytest
 import rdkit
 from gcms_data_analysis.main import Project
 
+test_dir: plib.Path = plib.Path(__file__).parent
 
+
+# testing name_to_properties
+name_to_properties_dir = test_dir / "data_name_to_properties"
+
+
+@pytest.fixture
+def dicts_classifications_codes_fractions():
+    ccf = pd.read_excel(
+        plib.Path(
+            name_to_properties_dir,
+            "classifications_codes_fractions.xlsx",
+        )
+    )
+    dict_class_to_code: dict[str, str] = dict(
+        zip(
+            ccf.classes.tolist(),
+            ccf.codes.tolist(),
+        )
+    )
+    dict_class_to_mass_fraction: dict[str, float] = dict(
+        zip(
+            ccf.classes.tolist(),
+            ccf.mfs.tolist(),
+        )
+    )
+    return dict_class_to_code, dict_class_to_mass_fraction
+
+
+@pytest.fixture
+def checked_compounds_properties():
+    properties = pd.read_excel(
+        plib.Path(
+            name_to_properties_dir,
+            "checked_compounds_properties_correct.xlsx",
+        ),
+        index_col="comp_name",
+    )
+    return properties
+
+
+# Project class testing
 @pytest.fixture
 def gcms() -> Project:
 
@@ -13,6 +55,7 @@ def gcms() -> Project:
         plib.Path(__file__).parent.parent, "tests/data_for_testing/"
     )
     Project.set_folder_path(folder_path)
+    Project.auto_save_to_excel(False)
     return Project()
 
 
@@ -202,44 +245,44 @@ def checked_list_of_all_deriv_compounds():
     ]
     return list_of_all_deriv_compounds
 
-@pytest.fixture
-def checked_compounds_properties():
-    compounds_properties = pd.DataFrame(
-        index=pd.Index(['tetradecanoic acid', 'oxacycloheptadecan-2-one', 'n-hexadecanoic acid', '9,12-octadecadienoic acid (z,z)-', 'oleic acid', 'n-decanoic acid', '2-butanone', '2-cyclopenten-1-one, 2-methyl-', 'trans-2-pentenoic acid', '2,5-hexanedione', '1-hexene, 4,5-dimethyl-', 'phenol', '2-methylcyclopent-2-en-1-one', '2,4,5-trichlorophenol', 'hexadecanoic acid', '(9z,12z)-octadeca-9,12-dienoic acid', '(z)-octadec-9-enoic acid'], name='comp_name'),
-        data={
-            'iupac_name': ['tetradecanoic acid', 'oxacycloheptadecan-2-one', 'hexadecanoic acid', '(9z,12z)-octadeca-9,12-dienoic acid', '(z)-octadec-9-enoic acid', 'decanoic acid', 'butan-2-one', '2-methylcyclopent-2-en-1-one', '(e)-pent-2-enoic acid', 'hexane-2,5-dione', '4,5-dimethylhex-1-ene', 'phenol', '2-methylcyclopent-2-en-1-one', '2,4,5-trichlorophenol', 'hexadecanoic acid', '(9z,12z)-octadeca-9,12-dienoic acid', '(z)-octadec-9-enoic acid'],
-            'molecular_formula': ['C14H28O2', 'C16H30O2', 'C16H32O2', 'C18H32O2', 'C18H34O2', 'C10H20O2', 'C4H8O', 'C6H8O', 'C5H8O2', 'C6H10O2', 'C8H16', 'C6H6O', 'C6H8O', 'C6H3Cl3O', 'C16H32O2', 'C18H32O2', 'C18H34O2'],
-            'canonical_smiles': ['CCCCCCCCCCCCCC(=O)O', 'C1CCCCCCCC(=O)OCCCCCCC1', 'CCCCCCCCCCCCCCCC(=O)O', 'CCCCCC=CCC=CCCCCCCCC(=O)O', 'CCCCCCCCC=CCCCCCCCC(=O)O', 'CCCCCCCCCC(=O)O', 'CCC(=O)C', 'CC1=CCCC1=O', 'CCC=CC(=O)O', 'CC(=O)CCC(=O)C', 'CC(C)C(C)CC=C', 'C1=CC=C(C=C1)O', 'CC1=CCCC1=O', 'C1=C(C(=CC(=C1Cl)Cl)Cl)O', 'CCCCCCCCCCCCCCCC(=O)O', 'CCCCCC=CCC=CCCCCCCCC(=O)O', 'CCCCCCCCC=CCCCCCCCC(=O)O'],
-            'molecular_weight': [228.37, 254.41, 256.42, 280.4, 282.5, 172.26, 72.11, 96.13, 100.12, 114.14, 112.21, 94.11, 96.13, 197.4, 256.42, 280.4, 282.5],
-            'xlogp': [5.3, 6.3, 6.4, 6.8, 6.5, 4.1, 0.3, 0.9, 1.0, -0.3, 3.5, 1.5, 0.9, 3.7, 6.4, 6.8, 6.5],
-            'el_C': [14, 16, 16, 18, 18, 10, 4, 6, 5, 6, 8, 6, 6, 6, 16, 18, 18],
-            'el_Cl': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0],
-            'el_H': [28, 30, 32, 32, 34, 20, 8, 8, 8, 10, 16, 6, 8, 3, 32, 32, 34],
-            'el_O': [2, 2, 2, 2, 2, 2, 1, 1, 2, 2, 0, 1, 1, 1, 2, 2, 2],
-            'el_mf_C': [0.7363226343214958, 0.7553791124562713, 0.7494579205990172, 0.7710342368045648, 0.7653026548672566, 0.6972599558806455, 0.6662598807377618, 0.7496723187350464, 0.5998302037554933, 0.6313825127036973, 0.8563229658675697, 0.765763468281798, 0.7496723187350464, 0.3650759878419453, 0.7494579205990172, 0.7710342368045648, 0.7653026548672566],
-            'el_mf_Cl': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.5387537993920973, 0.0, 0.0, 0.0],
-            'el_mf_H': [0.12358891272934273, 0.11886325223065132, 0.12579361984244597, 0.11503566333808846, 0.12131681415929203, 0.11703239289446186, 0.11182914990985994, 0.08388640382814938, 0.08054334798242109, 0.08831259856316805, 0.1437305053025577, 0.06426522154925088, 0.08388640382814938, 0.015319148936170212, 0.12579361984244597, 0.11503566333808846, 0.12131681415929203],
-            'el_mf_O': [0.1401147261023777, 0.12577335796548877, 0.12478745807659308, 0.11411554921540658, 0.11326725663716815, 0.18575409265064438, 0.22186936624601306, 0.16643087485696453, 0.31959648421893727, 0.28033993341510427, 0.0, 0.17000318775900541, 0.16643087485696453, 0.08104863221884498, 0.12478745807659308, 0.11411554921540658, 0.11326725663716815],
-            'fg_C-aliph': [13, 14, 15, 17, 17, 9, 1, 3, 4, 0, 8, 0, 3, 0, 15, 17, 17],
-            'fg_C-arom': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 6, 0, 0, 0],
-            'fg_Cl': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0],
-            'fg_alcohol': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0],
-            'fg_carboxyl': [1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1],
-            'fg_ester': [0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-            'fg_hetero_atoms': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0],
-            'fg_ketone': [0, 0, 0, 0, 0, 0, 1, 1, 0, 2, 0, 0, 1, 0, 0, 0, 0],
-            'fg_mf_C-aliph': [0.8029031834303979, 0.771895758814512, 0.8244793697839481, 0.8396398002853066, 0.8405345132743363, 0.7387147335423198, 0.20850090140063793, 0.4377509622386352, 0.5503395924890131, 0.0, 1.0000534711701274, 0.0, 0.4377509622386352, 0.0, 0.8244793697839481, 0.8396398002853066, 0.8405345132743363],
-            'fg_mf_C-arom': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.8193178195728402, 0.0, 0.3752887537993921, 0.0, 0.0, 0.0],
-            'fg_mf_Cl': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.5387537993920973, 0.0, 0.0, 0.0],
-            'fg_mf_alcohol': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1807140580172139, 0.0, 0.0861550151975684, 0.0, 0.0, 0.0],
-            'fg_mf_carboxyl': [0.19712308972281825, 0.0, 0.1755596287341081, 0.16054564907275323, 0.15935221238938055, 0.2613317078834321, 0.0, 0.0, 0.4496304434678386, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1755596287341081, 0.16054564907275323, 0.15935221238938055],
-            'fg_mf_ester': [0.0, 0.22811996383789943, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-            'fg_mf_hetero_atoms': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.5387537993920973, 0.0, 0.0, 0.0],
-            'fg_mf_ketone': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.7914574954929968, 0.5936960366170811, 0.0, 1.0000350446819695, 0.0, 0.0, 0.5936960366170811, 0.0, 0.0, 0.0, 0.0],
-            'fg_mf_total': [0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168],
-        }
-    )
-    return compounds_properties
+# @pytest.fixture
+# def checked_compounds_properties():
+#     compounds_properties = pd.DataFrame(
+#         index=pd.Index(['tetradecanoic acid', 'oxacycloheptadecan-2-one', 'n-hexadecanoic acid', '9,12-octadecadienoic acid (z,z)-', 'oleic acid', 'n-decanoic acid', '2-butanone', '2-cyclopenten-1-one, 2-methyl-', 'trans-2-pentenoic acid', '2,5-hexanedione', '1-hexene, 4,5-dimethyl-', 'phenol', '2-methylcyclopent-2-en-1-one', '2,4,5-trichlorophenol', 'hexadecanoic acid', '(9z,12z)-octadeca-9,12-dienoic acid', '(z)-octadec-9-enoic acid'], name='comp_name'),
+#         data={
+#             'iupac_name': ['tetradecanoic acid', 'oxacycloheptadecan-2-one', 'hexadecanoic acid', '(9z,12z)-octadeca-9,12-dienoic acid', '(z)-octadec-9-enoic acid', 'decanoic acid', 'butan-2-one', '2-methylcyclopent-2-en-1-one', '(e)-pent-2-enoic acid', 'hexane-2,5-dione', '4,5-dimethylhex-1-ene', 'phenol', '2-methylcyclopent-2-en-1-one', '2,4,5-trichlorophenol', 'hexadecanoic acid', '(9z,12z)-octadeca-9,12-dienoic acid', '(z)-octadec-9-enoic acid'],
+#             'molecular_formula': ['C14H28O2', 'C16H30O2', 'C16H32O2', 'C18H32O2', 'C18H34O2', 'C10H20O2', 'C4H8O', 'C6H8O', 'C5H8O2', 'C6H10O2', 'C8H16', 'C6H6O', 'C6H8O', 'C6H3Cl3O', 'C16H32O2', 'C18H32O2', 'C18H34O2'],
+#             'canonical_smiles': ['CCCCCCCCCCCCCC(=O)O', 'C1CCCCCCCC(=O)OCCCCCCC1', 'CCCCCCCCCCCCCCCC(=O)O', 'CCCCCC=CCC=CCCCCCCCC(=O)O', 'CCCCCCCCC=CCCCCCCCC(=O)O', 'CCCCCCCCCC(=O)O', 'CCC(=O)C', 'CC1=CCCC1=O', 'CCC=CC(=O)O', 'CC(=O)CCC(=O)C', 'CC(C)C(C)CC=C', 'C1=CC=C(C=C1)O', 'CC1=CCCC1=O', 'C1=C(C(=CC(=C1Cl)Cl)Cl)O', 'CCCCCCCCCCCCCCCC(=O)O', 'CCCCCC=CCC=CCCCCCCCC(=O)O', 'CCCCCCCCC=CCCCCCCCC(=O)O'],
+#             'molecular_weight': [228.37, 254.41, 256.42, 280.4, 282.5, 172.26, 72.11, 96.13, 100.12, 114.14, 112.21, 94.11, 96.13, 197.4, 256.42, 280.4, 282.5],
+#             'xlogp': [5.3, 6.3, 6.4, 6.8, 6.5, 4.1, 0.3, 0.9, 1.0, -0.3, 3.5, 1.5, 0.9, 3.7, 6.4, 6.8, 6.5],
+#             'el_C': [14, 16, 16, 18, 18, 10, 4, 6, 5, 6, 8, 6, 6, 6, 16, 18, 18],
+#             'el_Cl': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0],
+#             'el_H': [28, 30, 32, 32, 34, 20, 8, 8, 8, 10, 16, 6, 8, 3, 32, 32, 34],
+#             'el_O': [2, 2, 2, 2, 2, 2, 1, 1, 2, 2, 0, 1, 1, 1, 2, 2, 2],
+#             'el_mf_C': [0.7363226343214958, 0.7553791124562713, 0.7494579205990172, 0.7710342368045648, 0.7653026548672566, 0.6972599558806455, 0.6662598807377618, 0.7496723187350464, 0.5998302037554933, 0.6313825127036973, 0.8563229658675697, 0.765763468281798, 0.7496723187350464, 0.3650759878419453, 0.7494579205990172, 0.7710342368045648, 0.7653026548672566],
+#             'el_mf_Cl': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.5387537993920973, 0.0, 0.0, 0.0],
+#             'el_mf_H': [0.12358891272934273, 0.11886325223065132, 0.12579361984244597, 0.11503566333808846, 0.12131681415929203, 0.11703239289446186, 0.11182914990985994, 0.08388640382814938, 0.08054334798242109, 0.08831259856316805, 0.1437305053025577, 0.06426522154925088, 0.08388640382814938, 0.015319148936170212, 0.12579361984244597, 0.11503566333808846, 0.12131681415929203],
+#             'el_mf_O': [0.1401147261023777, 0.12577335796548877, 0.12478745807659308, 0.11411554921540658, 0.11326725663716815, 0.18575409265064438, 0.22186936624601306, 0.16643087485696453, 0.31959648421893727, 0.28033993341510427, 0.0, 0.17000318775900541, 0.16643087485696453, 0.08104863221884498, 0.12478745807659308, 0.11411554921540658, 0.11326725663716815],
+#             'fg_C-aliph': [13, 14, 15, 17, 17, 9, 1, 3, 4, 0, 8, 0, 3, 0, 15, 17, 17],
+#             'fg_C-arom': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 6, 0, 0, 0],
+#             'fg_Cl': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0],
+#             'fg_alcohol': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0],
+#             'fg_carboxyl': [1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 1],
+#             'fg_ester': [0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+#             'fg_hetero_atoms': [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0],
+#             'fg_ketone': [0, 0, 0, 0, 0, 0, 1, 1, 0, 2, 0, 0, 1, 0, 0, 0, 0],
+#             'fg_mf_C-aliph': [0.8029031834303979, 0.771895758814512, 0.8244793697839481, 0.8396398002853066, 0.8405345132743363, 0.7387147335423198, 0.20850090140063793, 0.4377509622386352, 0.5503395924890131, 0.0, 1.0000534711701274, 0.0, 0.4377509622386352, 0.0, 0.8244793697839481, 0.8396398002853066, 0.8405345132743363],
+#             'fg_mf_C-arom': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.8193178195728402, 0.0, 0.3752887537993921, 0.0, 0.0, 0.0],
+#             'fg_mf_Cl': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.5387537993920973, 0.0, 0.0, 0.0],
+#             'fg_mf_alcohol': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1807140580172139, 0.0, 0.0861550151975684, 0.0, 0.0, 0.0],
+#             'fg_mf_carboxyl': [0.19712308972281825, 0.0, 0.1755596287341081, 0.16054564907275323, 0.15935221238938055, 0.2613317078834321, 0.0, 0.0, 0.4496304434678386, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1755596287341081, 0.16054564907275323, 0.15935221238938055],
+#             'fg_mf_ester': [0.0, 0.22811996383789943, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
+#             'fg_mf_hetero_atoms': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.5387537993920973, 0.0, 0.0, 0.0],
+#             'fg_mf_ketone': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.7914574954929968, 0.5936960366170811, 0.0, 1.0000350446819695, 0.0, 0.0, 0.5936960366170811, 0.0, 0.0, 0.0, 0.0],
+#             'fg_mf_total': [0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168, 0.9998867256637168],
+#         }
+#     )
+#     return compounds_properties
 
 @pytest.fixture
 def checked_deriv_compounds_properties():
